pyproject.toml
[project]
name = "YA_MCPServer_Template"
version = "0.1.0"
description = "YA MCP Server Template"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.3",
    "art>=6.5",
    "black>=25.9.0",
    "colorlog>=6.10.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.14.0",
    "pyyaml>=6.0.2",
    "ruff>=0.14.4",
]
